@@ -1,6 +1,7 @@
 #pragma once
 
 #include "docstr.hpp"
+#include "misc/helpers.hpp"
 #include <nanobind/eigen/dense.h>
 #include <nanobind/nanobind.h>
 #include <nanobind/stl/tuple.h>
@@ -9,6 +10,19 @@
 namespace nb = nanobind;
 using namespace nb::literals;
 using namespace vinecopulib;
+
+// Wrapper function to call the Python bicop_plot function
+inline void
+bicop_plot_wrapper(const Bicop& cop,
+                   const std::string& type,
+                   const std::string& margin_type,
+                   nb::object xylim,
+                   nb::object grid_size)
+{
+  auto mod = nb::module_::import_("pyvinecopulib._python_helpers.bicop");
+  auto bicop_plot = mod.attr("bicop_plot");
+  bicop_plot(nb::cast(cop), type, margin_type, xylim, grid_size);
+}
 
 // Factory function to create a Bicop from family, rotation, parameters, and
 // variable types
@@ -106,7 +120,7 @@ Alternatives to instantiate bivariate copulas are:
                  "The copula rotation.")
     .def_prop_rw(
       "parameters",
-      [](Bicop& self) { return nb::cast(self.get_parameters()); },
+      &Bicop::get_parameters,
       [](Bicop& self, const nb::DRef<Eigen::MatrixXd>& parameters) {
         self.set_parameters(parameters);
       },
@@ -141,9 +155,14 @@ Alternatives to instantiate bivariate copulas are:
          "u"_a = Eigen::Matrix<double, Eigen::Dynamic, 2>(),
          "psi0"_a = 0.9,
          bicop_doc.mbic.doc)
-    .def("__repr__",
-         [](const Bicop& cop) { return "<pyvinecopulib.Bicop>\n" + cop.str(); })
-    .def("str", &Bicop::str, bicop_doc.str.doc)
+    .def(
+      "__repr__",
+      [](const Bicop& cop) { return "<pyvinecopulib.Bicop> " + cop.str(); },
+      bicop_doc.str.doc)
+    .def(
+      "__str__",
+      [](const Bicop& cop) { return "<pyvinecopulib.Bicop> " + cop.str(); },
+      bicop_doc.str.doc)
     .def("parameters_to_tau",
          &Bicop::parameters_to_tau,
          "parameters"_a,
@@ -152,18 +171,12 @@ Alternatives to instantiate bivariate copulas are:
          &Bicop::tau_to_parameters,
          "tau"_a,
          bicop_doc.tau_to_parameters.doc)
-    .def_prop_ro(
-      "parameters_lower_bounds",
-      [](const Bicop& self) {
-        return nb::cast(self.get_parameters_lower_bounds());
-      },
-      bicop_doc.get_parameters_lower_bounds.doc)
-    .def_prop_ro(
-      "parameters_upper_bounds",
-      [](const Bicop& self) {
-        return nb::cast(self.get_parameters_upper_bounds());
-      },
-      bicop_doc.get_parameters_upper_bounds.doc)
+    .def_prop_ro("parameters_lower_bounds",
+                 &Bicop::get_parameters_lower_bounds,
+                 bicop_doc.get_parameters_lower_bounds.doc)
+    .def_prop_ro("parameters_upper_bounds",
+                 &Bicop::get_parameters_upper_bounds,
+                 bicop_doc.get_parameters_upper_bounds.doc)
     .def("pdf", &Bicop::pdf, "u"_a, bicop_doc.pdf.doc)
     .def("cdf", &Bicop::cdf, "u"_a, bicop_doc.cdf.doc)
     .def("hfunc1", &Bicop::hfunc1, "u"_a, bicop_doc.hfunc1.doc)
@@ -188,27 +201,14 @@ Alternatives to instantiate bivariate copulas are:
          bicop_doc.select.doc)
     .def(
       "plot",
-      [](const Bicop& cop,
-         const std::string& type = "surface",
-         const std::string& margin_type = "unif",
-         nb::object xylim = nb::none(),
-         nb::object grid_size = nb::none()) {
-        auto python_helpers_plotting =
-          nb::module_::import_("pyvinecopulib._python_helpers.bicop");
-
-        // Import the Python plotting function
-        nb::object bicop_plot = python_helpers_plotting.attr("bicop_plot");
-
-        // Call the Python function with the provided arguments
-        bicop_plot(nb::cast(cop), type, margin_type, xylim, grid_size);
-      },
+      &bicop_plot_wrapper,
       "type"_a = "surface",
       "margin_type"_a = "unif",
       "xylim"_a = nb::none(),
       "grid_size"_a = nb::none(),
-      nb::cast<std::string>(
-        nb::module_::import_("pyvinecopulib._python_helpers.bicop")
-          .attr("BICOP_PLOT_DOC"))
+      python_doc_helper("pyvinecopulib._python_helpers.bicop",
+                        "BICOP_PLOT_DOC",
+                        "Plot the bivariate copula (extended doc unavailable) ")
         .c_str())
     .def("__getstate__",
          [](const Bicop& cop) {
