@@ -1,7 +1,10 @@
+from typing import Any
+
 import numpy as np
+from numpy.typing import NDArray
 
 
-def random_data(d=5, n=1000):
+def random_data(d: int = 5, n: int = 1000) -> NDArray[np.float64]:
   # Simulate some data
   np.random.seed(1234)  # seed for the random generator
   mean = np.random.normal(size=d)  # mean vector
@@ -11,30 +14,32 @@ def random_data(d=5, n=1000):
   return x
 
 
-def compare_properties(obj1, obj2, attrs, subclass=False):
+def compare_properties(
+  obj1: Any, obj2: Any, attrs: list[str], subclass: bool = False
+) -> None:
   if subclass:
-    assert issubclass(
-      type(obj1), type(obj2)
-    ), "Objects must be of the same type"
+    assert issubclass(type(obj1), type(obj2)), (
+      "Objects must be of the same type"
+    )
   else:
     assert type(obj1) is type(obj2), "Objects must be of the same type"
   for attr in attrs:
     val1 = getattr(obj1, attr)
     val2 = getattr(obj2, attr)
     if isinstance(val1, np.ndarray):
-      assert isinstance(val2, np.ndarray) and np.array_equal(
-        val1, val2
-      ), f"Mismatch in {attr}: {val1} != {val2}"
+      assert isinstance(val2, np.ndarray) and np.array_equal(val1, val2), (
+        f"Mismatch in {attr}: {val1} != {val2}"
+      )
     else:
       assert val1 == val2, f"Mismatch in {attr}: {val1} != {val2}"
 
 
-def compare_bicop(cop1, cop2):
+def compare_bicop(cop1: Any, cop2: Any) -> None:
   attrs = ["family", "rotation", "parameters", "var_types"]
   compare_properties(cop1, cop2, attrs)
 
 
-def compare_vinecop(cop1, cop2):
+def compare_vinecop(cop1: Any, cop2: Any) -> None:
   attrs = ["dim", "trunc_lvl", "order", "matrix"]
   compare_properties(cop1, cop2, attrs)
 
@@ -47,6 +52,8 @@ def compare_vinecop(cop1, cop2):
       compare_bicop(bicop1, bicop2)
 
 
-def compare_rvinestructure(struct1, struct2, subclass=False):
+def compare_rvinestructure(
+  struct1: Any, struct2: Any, subclass: bool = False
+) -> None:
   attrs = ["dim", "order", "trunc_lvl", "matrix"]
   compare_properties(struct1, struct2, attrs, subclass)
